Start package TTF2 (fresh new parser)
Just a start of a new parser for the true type font file. The idea is
to simplify the management of all the allocations by bundling everything
into an arena and to flatten out the tables into an easy to access
enum array.

Also introduced is the concept of a 'Reader' just to help with OOB
and debugging.

Tests are provided to cover all the functionality right now (which is
just gathering all the tables in the file).