Use glyph array from ttf2 in raster
Start of glooping over ttf2 stuff into ttf. I though the easiest
thing to tackle would be the loading/caching of extracted glyphs.

It is a little dodge since I am tacking on this parser onto the
existing one but I think it's the easiest way to combine things
over slowly.

The glyphs are initially not extracted at all and we just reserve
a slice for quick lookups. As discussed before we would likely
in the future provide a user supplied range of codepoints to
pre extract glyphs at load time.