Add the chunky funky hinter
I have tested the runic hinter in my own project so it looks pretty
good to go.

Alongside the hinter I have added a 'memory' package which just
has some utils for managing memory. The first thing added is a scratch
arena implementation. It just gives you an arena that does not conflict
with the provided allocators. Secondly I have added a 'make_multi'
procedure. This is just some auto generated code that lets you allocate
multiple slices/structs in a single call. Really it's used to reduce
the number of errors needed to be handled by the user (rather than
an optimisation of any kind).

After this patch I will work on getting rid of ttf2 :)