Add private arena per font
This patch adds a private arena for the font where the font
itself is located and where tables can allocate whatever they
want from. This removes all cases of needing a 'destroy' proc
so I have removed it for now.

The arena uses some heuristics to guess what size it should
allocate but it can be overriden by the user.